config: Import Product model
Include the import statement for the `Product` model from the current application's models module. This import is necessary for utilizing the Product model in other parts of the codebase. This commit ensures that the Product model is properly imported and available for use in relevant modules.

feat(admin): Customize admin panel for Product

Create the `ProductAdmin` class to customize the admin panel for the `Product` model. Specify the fields to be displayed in the list view, including product name, price, stock, category, modified date, and availability. Additionally, set up prepopulated fields for the slug based on the product name. This commit enhances the visual representation and management of products in the Django admin panel.