style: Define rating choices and category model
Introduced 'RATING' choices for rating stars and implemented the 'Category' model. The model includes a 'cid' field with ShortUUID for a unique identifier, a 'title' field for the category title, and an 'image' field for category images. Also, added a 'category_image' method for displaying images in the admin panel using 'mark_safe'.