style: Define Tags and Vendor models
Introduced the 'Tags' and 'Vendor' models. The 'Vendor' model includes fields such as 'vid' with ShortUUID for a unique identifier, 'title' for the vendor title, 'image' for vendor images, and various other fields for vendor details. Added a 'vendor_image' method for displaying images in the admin panel using 'mark_safe'.