style: Update category display template
In the template, modified the way products associated with each category are counted and displayed. Instead of using `c.product_set.all|length`, it suggests using the related name `c.category.count` since the related name is set in the model. The change enhances the efficiency and readability of the template.