feat: Add ForeignKey relationship for product images
In this commit, a ForeignKey relationship is added to the Product model to establish a connection between the Product and its associated images. The related_name is set to "p_images" to allow easy access to the images associated with a product. This ForeignKey field allows each product to have multiple images associated with it.